+++ languages/de/tl_module.php
@@ -53,6 +53,9 @@ $arrLang['share_pdfUsername'][1] = 'Falls Sie Daten (Bilder,CSS) von einer Seite
 $arrLang['share_pdfPassword'][0] = 'Authentifizierungspassword';
 $arrLang['share_pdfPassword'][1] = 'Siehe Authentifizierungsbenutzername.';
 
+$arrLang['share_mailtoSubject'][0] = 'Betreff';
+$arrLang['share_mailtoSubject'][1] = 'Geben Sie hier den Betreff ein, der dem mailto-Link hinzugefügt werden soll. Wenn Sie nichts eingeben, wird der Titel der geteilten Entität genutzt. Sie können im Text mit "%s" auch den Titel der geteilten Entität referenzieren.';
+
 /**
  * Legends
  */
@@ -66,3 +69,4 @@ $arrLang['references']['share_buttons']['printButton'] = 'Beitrag Drucken';
 $arrLang['references']['share_buttons']['facebook']    = 'Auf Facebook teilen';
 $arrLang['references']['share_buttons']['twitter']     = 'Auf Twitter teilen';
 $arrLang['references']['share_buttons']['gplus']       = 'Auf Google+ teilen';
+$arrLang['references']['share_buttons']['mailto']      = 'Per E-Mail teilen';